requirements.txt
numpy
scipy
pillow
matplotlib
tqdm
